api/requirements.txt
fastapi==0.45.0
FastAPI-SQLAlchemy==0.1.0
pydantic==1.9.2
alembic==1.7.7
psycopg==3.0.16
psycopg2-binary==2.9.3
uvicorn==0.16.0
python-dotenv==0.20.0
contextvars==2.4